rename util to json & remove lets-plot mention
