+++ settings.gradle.kts
@@ -47,7 +47,7 @@ include("geotools-batik")
 include("dokka")
 include("jupyter")
 include("geotools-jupyter")
-include("util")
+include("json")
 
 project(":demo-common").projectDir = File("./demo/demo-common")
 project(":jvm-javafx").projectDir = File("./demo/jvm-javafx")
@@ -63,4 +63,4 @@ project(":dokka").projectDir = File("./docs/dokka")
 project(":jupyter").projectDir = File("./toolkit/jupyter")
 project(":geotools-jupyter").projectDir = File("./toolkit/geotools-jupyter")
 
-project(":util").projectDir = File("./toolkit/util")
+project(":json").projectDir = File("./toolkit/json")
